+++ src/main/java/kitchenpos/menu/domain/Menu.java
@@ -1,7 +1,5 @@
 package kitchenpos.menu.domain;
 
-import kitchenpos.menuGroup.domain.MenuGroup;
-
 import javax.persistence.*;
 import java.math.BigDecimal;
 import java.util.List;
