+++ src/main/java/kitchenpos/menu/domain/MenuGroup.java
@@ -1,4 +1,4 @@
-package kitchenpos.menuGroup.domain;
+package kitchenpos.menu.domain;
 
 import javax.persistence.*;
