+++ src/main/java/kitchenpos/menu/dto/MenuRequest.java
@@ -1,8 +1,7 @@
 package kitchenpos.menu.dto;
 
 import kitchenpos.menu.domain.Menu;
-import kitchenpos.menu.domain.MenuProduct;
-import kitchenpos.menuGroup.domain.MenuGroup;
+import kitchenpos.menu.domain.MenuGroup;
 
 import java.math.BigDecimal;
 import java.util.List;
