+++ src/test/java/kitchenpos/fixture/MenuFixture.java
@@ -1,11 +1,9 @@
 package kitchenpos.fixture;
 
 import kitchenpos.menu.domain.Menu;
-import kitchenpos.menu.domain.MenuProduct;
-import kitchenpos.menuGroup.domain.MenuGroup;
+import kitchenpos.menu.domain.MenuGroup;
 
 import java.math.BigDecimal;
-import java.util.List;
 
 public class MenuFixture {
