+++ src/test/java/kitchenpos/menu/application/MenuServiceTest.java
@@ -6,9 +6,9 @@ import kitchenpos.menu.domain.*;
 import kitchenpos.menu.dto.MenuProductRequest;
 import kitchenpos.menu.dto.MenuRequest;
 import kitchenpos.menu.dto.MenuResponse;
-import kitchenpos.menuGroup.domain.MenuGroup;
+import kitchenpos.menu.domain.MenuGroup;
 import kitchenpos.product.domain.Product;
-import kitchenpos.menuGroup.domain.MenuGroupRepository;
+import kitchenpos.menu.domain.MenuGroupRepository;
 import kitchenpos.fixture.MenuGroupFixture;
 import kitchenpos.fixture.MenuFixture;
 import kitchenpos.product.domain.ProductRepository;
